+++ src/jsonyx/test/test_jsonyx.py
@@ -27,7 +27,8 @@ def test_duplicate_key(json: ModuleType) -> None:
     assert isinstance(key, str)
     assert not str(key)
     assert key != ""  # noqa: PLC1901
-    assert hash(key) == id(key)
+    assert not key == ""  # noqa: SIM201, PLC1901
+    assert hash(key) == object.__hash__(key)  # noqa: PLC2801
 
 
 @pytest.mark.parametrize(("s", "encoding"), [