+++ docs/source/changelog.rst
@@ -22,6 +22,7 @@ jsonyx 2.0.0 (unreleased)
 - Added :func:`jsonyx.Manipulator`
 - Changed error for big integers to :exc:`jsonyx.JSONSyntaxError`
 - Fixed line comment detection
+- Fixed equality of :class:`jsonyx.DuplicateKey`
 - Fixed typo in error message
 - Improved documentation
 - Made :class:`tuple` JSON serializable